Move impl to Compiler
